Don't overwrite if note exists on new

+++ lib/abulafia/commands/new.rb
@@ -9,9 +9,12 @@ module Abulafia
 
         name = args.join(' ')
         path = to_slug(name)
-        cfg.repo.open(path, 'w') do |f|
-          f.puts "# #{name}\n@inbox\n\n"
-          f.puts "created_at { #{cfg.time.stamp} }\n\n"
+
+        if !cfg.repo.exist?(path)
+          cfg.repo.open(path, 'w') do |f|
+            f.puts "# #{name}\n@inbox\n\n"
+            f.puts "created_at { #{cfg.time.stamp} }\n\n"
+          end
         end
         cfg.editor.open(cfg.repo.full_path(path))
       end
